Fix --cacert passing blank arg
As of curl 8.6.0, curl fails if a blank arg is passed.  https://github.com/curl/curl/pull/12620

Default behavior of tosca_execution_client.sh is to pass a blank --cacert to curl if --cacertificate is not specified when running the execution client script.

This modifies the execution client to not pass --cacert at all if --cacertificate is not specified when running the execution client.

If --cacertificate is specified, curl is passed "--cacert " followed by the value of the --cacertificate flag specified when running the execution client.